Move viewmodel out to its own file
