+++ app/src/main/java/com/neaniesoft/vermilion/app/App.kt
@@ -1,7 +1,5 @@
 package com.neaniesoft.vermilion.app
 
-import android.os.Bundle
-import android.util.Log
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
@@ -15,9 +13,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import androidx.navigation.NavDestination
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.plusAssign
@@ -28,21 +23,8 @@ import com.google.accompanist.navigation.material.rememberBottomSheetNavigator
 import com.neaniesoft.vermilion.app.customtabs.CustomTabNavigator
 import com.neaniesoft.vermilion.tabs.adapters.driving.ui.ActiveTab
 import com.neaniesoft.vermilion.tabs.adapters.driving.ui.TabBottomBar
-import com.neaniesoft.vermilion.tabs.domain.TabSupervisor
-import com.neaniesoft.vermilion.tabs.domain.entities.ParentId
-import com.neaniesoft.vermilion.tabs.domain.entities.TabState
-import com.neaniesoft.vermilion.tabs.domain.entities.TabType
 import com.neaniesoft.vermilion.ui.theme.VermilionTheme
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
 import java.time.Clock
-import javax.inject.Inject
 
 @ExperimentalPagingApi
 @ExperimentalMaterialNavigationApi
@@ -61,7 +43,7 @@ fun VermilionApp(
         navController.navigatorProvider += bottomSheetNavigator
         navController.navigatorProvider += customTabNavigator
 
-        navController.addOnDestinationChangedListener { controller, destination, arguments ->
+        navController.addOnDestinationChangedListener { _, destination, arguments ->
             viewModel.onNavigationEvent(destination, arguments)
         }
 
@@ -87,7 +69,7 @@ fun VermilionApp(
 
         LaunchedEffect(key1 = activeTab, block = {
             viewModel.clearTabFromBackstackEvents.collect { clearTab ->
-                @Suppress("UnnecessaryVariable")
+                @Suppress("UnnecessaryVariable") // can't inline because it has a custom getter
                 val currentTab = activeTab
                 if (currentTab is ActiveTab.Tab && currentTab.id == clearTab.id) {
                     navController.popBackStack(VermilionScreen.Home.name, false)
@@ -129,67 +111,3 @@ fun VermilionApp(
     }
 }
 
-@HiltViewModel
-class VermilionAppViewModel @Inject constructor(
-    private val tabSupervisor: TabSupervisor
-) : ViewModel() {
-    val tabs = tabSupervisor.currentTabs
-
-    private val _routeEvents = MutableSharedFlow<String>()
-    val routeEvents: SharedFlow<String> = _routeEvents.asSharedFlow()
-
-    private val _clearTabFromBackstackEvents = MutableSharedFlow<TabState>()
-    val clearTabFromBackstackEvents = _clearTabFromBackstackEvents.asSharedFlow()
-
-    private val _activeTab = MutableStateFlow<ActiveTab>(ActiveTab.None)
-    val activeTab = _activeTab.asStateFlow()
-
-    fun onNavigationEvent(destination: NavDestination, args: Bundle?) {
-        val route = destination.route
-        Log.d("VermilionAppViewModel", "Route: $route; args: $args")
-        if (route != null) {
-            when {
-                route.startsWith(VermilionScreen.PostDetails.name) -> {
-                    val id =
-                        requireNotNull(args?.getString("id")) { "Received a post details route with no id" }
-                    viewModelScope.launch(Dispatchers.IO) {
-                        val tab = tabSupervisor.addNewPostDetailsTabIfNotExists(ParentId(id))
-                        _activeTab.emit(ActiveTab.Tab(tab.id))
-                    }
-                }
-                route.startsWith(VermilionScreen.Home.name) -> {
-                    viewModelScope.launch { _activeTab.emit(ActiveTab.Home) }
-                }
-                else -> {
-                    viewModelScope.launch { _activeTab.emit(ActiveTab.None) }
-                }
-            }
-        }
-    }
-
-    fun onTabClicked(tab: TabState) {
-        emitRouteEvent(tab.type, tab.parentId)
-    }
-
-    private fun emitRouteEvent(type: TabType, parentId: ParentId) {
-        if (type == TabType.POST_DETAILS) {
-            val route = "${VermilionScreen.PostDetails}/${parentId.value}"
-            viewModelScope.launch { _routeEvents.emit(route) }
-        }
-    }
-
-    fun onTabCloseClicked(tab: TabState) {
-        viewModelScope.launch {
-            tabSupervisor.removeTab(tab)
-            _clearTabFromBackstackEvents.emit(tab)
-        }
-    }
-
-    fun onHomeButtonClicked() {
-        viewModelScope.launch { _routeEvents.emit(VermilionScreen.Home.name) }
-    }
-
-    fun onUserButtonClicked() {
-        viewModelScope.launch { _routeEvents.emit(VermilionScreen.MyAccount.name) }
-    }
-}
